refactor: Workbook도메인의 비즈니스로직과 이벤트로직 분리
